requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
google-generative-ai>=0.15.0
